feat(recipe): Tauri CRUD commands
Wire six recipe commands (list/get/create/update/delete/restore) through
the manor-app Tauri layer, mirroring the ledger::commands Db-state pattern.
Fix three pre-existing clippy lints (map_identity, useless_conversion,
unnecessary_closure) that blocked -D warnings on manor-app.

+++ crates/app/src/ledger/bank_commands.rs
@@ -508,7 +508,7 @@ pub async fn ledger_bank_autocat_pending(
                 rusqlite::params![cat_id, tx_id],
             )
             .unwrap_or(0);
-        updated += rows as usize;
+        updated += rows;
     }
     Ok(updated)
 }
